proj2/backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
email-validator==2.2.0

